--- a/test/JDBC/expected/BABEL-SP_FKEYS.out
+++ b/test/JDBC/expected/BABEL-SP_FKEYS.out
@@ -1,3 +1,4 @@
+
 create database db1
 go
 use db1