--- a/test/JDBC/input/BABEL-SP_FKEYS-vu-verify.sql
+++ b/test/JDBC/input/BABEL-SP_FKEYS-vu-verify.sql
@@ -1,4 +1,3 @@
--- sla 2050000
 use babel_sp_fkeys_vu_prepare_db1
 go