--- a/test/JDBC/input/BABEL-SP_FKEYS.sql
+++ b/test/JDBC/input/BABEL-SP_FKEYS.sql
@@ -1,4 +1,4 @@
--- sla 1000000
+
 create database db1
 go
 use db1