--- a/test/JDBC/upgrade/13_6/schedule
+++ b/test/JDBC/upgrade/13_6/schedule
@@ -301,3 +301,5 @@ AVG-Aggregate-common
 AVG-Aggregate-Dep-before-15-2-or-14-7
 bbf_view_def-before-14_5
 BABEL_OBJECT_NAME
+BABEL-SP_FKEYS
+BABEL-SP_FKEYS-dep