--- a/test/JDBC/upgrade/13_9/schedule
+++ b/test/JDBC/upgrade/13_9/schedule
@@ -299,3 +299,5 @@ AVG-Aggregate-common
 AVG-Aggregate-Dep-before-15-2-or-14-7
 bbf_view_def-before-14_5
 BABEL_OBJECT_NAME
+BABEL-SP_FKEYS
+BABEL-SP_FKEYS-dep